requirements.txt
flask==3.1.0
flask-login==0.6.3
flask-limiter==3.8.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
requests==2.33.1
psycopg2-binary
